refactor: remove redundant search helpers
