+++ handlers/news/helpers.go
@@ -3,14 +3,8 @@ package news
 import (
 	"bytes"
 	"context"
-	"database/sql"
-	"errors"
 	"fmt"
-	"log"
-	"net/http"
-	"strings"
 	"text/template"
-	"unicode"
 
 	hcommon "github.com/arran4/goa4web/handlers/common"
 	db "github.com/arran4/goa4web/internal/db"
@@ -106,62 +100,6 @@ func notifyChange(ctx context.Context, provider email.Provider, emailAddr string
 		}
 	}
 	return nil
-}
-
-func breakupTextToWords(input string) []string {
-	var tokens []string
-	startIndex := -1
-
-	for i, char := range input {
-		if unicode.IsLetter(char) || unicode.IsDigit(char) || strings.ContainsRune("'-", char) {
-			if startIndex == -1 {
-				startIndex = i
-			}
-		} else if startIndex != -1 {
-			tokens = append(tokens, input[startIndex:i])
-			startIndex = -1
-		}
-	}
-	if startIndex != -1 {
-		tokens = append(tokens, input[startIndex:])
-	}
-	return tokens
-}
-
-func SearchWordIdsFromTextLocal(w http.ResponseWriter, r *http.Request, text string, queries *db.Queries) ([]int64, bool) {
-	words := map[string]int32{}
-	for _, word := range breakupTextToWords(text) {
-		words[strings.ToLower(word)] = 0
-	}
-	wordIds := make([]int64, 0, len(words))
-	for word := range words {
-		id, err := queries.CreateSearchWord(r.Context(), strings.ToLower(word))
-		if err != nil {
-			log.Printf("Error: createSearchWord: %s", err)
-			http.Redirect(w, r, "?error="+err.Error(), http.StatusTemporaryRedirect)
-			return nil, true
-		}
-		wordIds = append(wordIds, id)
-	}
-	return wordIds, false
-}
-
-func InsertWordsToForumSearchLocal(w http.ResponseWriter, r *http.Request, wordIds []int64, queries *db.Queries, cid int64) bool {
-	for _, wid := range wordIds {
-		if err := queries.AddToForumCommentSearch(r.Context(), db.AddToForumCommentSearchParams{
-			CommentsIdcomments:             int32(cid),
-			SearchwordlistIdsearchwordlist: int32(wid),
-		}); err != nil {
-			switch {
-			case errors.Is(err, sql.ErrNoRows):
-			default:
-				log.Printf("Error: addToForumCommentSearch: %s", err)
-				http.Redirect(w, r, "?error="+err.Error(), http.StatusTemporaryRedirect)
-				return true
-			}
-		}
-	}
-	return false
 }
 
 func PostUpdateLocal(ctx context.Context, q *db.Queries, threadID, topicID int32) error {
